fix: 필요없는 부분 제거 및 댓글 엔드포인트 위치 수정
