+++ backend/routes/community.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from fastapi import APIRouter, Depends, HTTPException, Query
+from fastapi import APIRouter, Depends, HTTPException, Query, Path
 from sqlalchemy.orm import Session
 from sqlalchemy import asc, desc, func, select
 from typing import List, Optional
@@ -661,3 +661,42 @@ def get_comments_for_post(db: Session, post_id: int):
         .order_by(models.Comment.created_at.desc())
         .all()
     )
+
+@router.delete("/comment/remove/{comment_id}",
+    response_model=dict,
+    summary="댓글 삭제",
+    description="""
+    지정된 ID의 댓글을 삭제합니다.
+    
+    삭제된 댓글은 복구할 수 없으며, 해당 댓글과 관련된 모든 데이터가 함께 삭제됩니다.
+    """,
+    response_description="삭제 완료 메시지",
+    responses={
+        200: {
+            "description": "댓글이 성공적으로 삭제됨",
+            "content": {
+                "application/json": {
+                    "example": {"message": "댓글이 삭제되었습니다."}
+                }
+            }
+        },
+        404: {
+            "description": "댓글을 찾을 수 없음",
+            "content": {
+                "application/json": {
+                    "example": {"detail": "댓글을 찾을 수 없습니다."}
+                }
+            }
+        }
+    }
+)
+def delete_comment(
+    comment_id: int = Path(..., description="삭제할 댓글의 ID", example=1),
+    db: Session = Depends(get_db)
+):
+    comment = db.query(models.Comment).filter(models.Comment.comment_id == comment_id).first()
+    if not comment:
+        raise HTTPException(status_code=404, detail="댓글을 찾을 수 없습니다.")
+    db.delete(comment)
+    db.commit()
+    return {"message": "댓글이 삭제되었습니다."}
